requirements.txt
# ============================================================
# Multi-Agent Literature Miner — Pinned Dependencies
# Generated: 2026-02-13  |  Python 3.x  |  Windows
# ============================================================
# Install:  pip install -r requirements.txt
# Reset:    pip freeze | xargs pip uninstall -y && pip install -r requirements.txt
# ============================================================

# --- LangChain Ecosystem (v0.3+ modular architecture) -------
langchain==1.2.10
langchain-core==1.2.12
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-google-genai==4.2.0
langchain-huggingface==1.2.0
sentence-transformers==5.2.2

# --- Vector Store -------------------------------------------
faiss-cpu==1.13.2

# --- Google GenAI SDK ---------------------------------------
google-genai==1.63.0

# --- BioInformatics -----------------------------------------
biopython==1.86

# --- Data & Visualization -----------------------------------
pandas==2.3.3
networkx==3.4.2
matplotlib==3.10.8
scipy==1.15.3

# --- Web UI -------------------------------------------------
gradio==6.5.1

# --- HTTP ---------------------------------------------------
requests==2.32.5

# --- Environment --------------------------------------------
python-dotenv==1.2.1
